+++ packages/types-codec/src/base/Option.spec.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { TypeRegistry } from '@polkadot/types';
-import { bool, Bytes, Option, Text, U32 } from '@polkadot/types-codec';
+import { bool, Bytes, Null, Option, Text, U32 } from '@polkadot/types-codec';
 
 const registry = new TypeRegistry();
 
@@ -27,6 +27,16 @@ describe('Option', (): void => {
     expect(new Option(registry, Text, null).isNone).toBe(true);
     expect(new Option(registry, Text, 'test').isNone).toBe(false);
     expect(new Option(registry, Text, '0x').isNone).toBe(true);
+  });
+
+  it('can wrap an Option<Null>', (): void => {
+    const test = new Option(registry, Null, new Null(registry));
+
+    expect(test.isSome).toBe(true);
+    expect(test.isNone).toBe(false);
+    expect(test.isEmpty).toBe(false);
+    expect(test.toU8a()).toEqual(new Uint8Array([1]));
+    expect(test.unwrap().toHex()).toEqual('0x');
   });
 
   it('correctly handles booleans', (): void => {